--- src/main/java/ListCommand.java
@@ -1,7 +1,0 @@
-public class ListCommand extends Command{
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        taskList.list();
-    }
-}
