--- src/main/java/ListOngoingCommand.java
@@ -1,7 +1,0 @@
-public class ListOngoingCommand extends Command{
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        taskList.listOngoing();
-    }
-}
