--- src/main/java/ListOverdueCommand.java
@@ -1,7 +1,0 @@
-public class ListOverdueCommand extends Command{
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        taskList.listOverdue();
-    }
-}
