--- src/main/java/ListUpcomingCommand.java
@@ -1,7 +1,0 @@
-public class ListUpcomingCommand extends Command{
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        taskList.listUpcoming();
-    }
-}
