+++ src/main/java/jamal/command/DeadlineTaskCommand.java
@@ -1,3 +1,11 @@
+package jamal.command;
+
+import jamal.task.Deadline;
+import jamal.task.Task;
+import jamal.task.TaskList;
+import jamal.ui.Ui;
+import jamal.util.Storage;
+
 public class DeadlineTaskCommand extends Command {
 
     protected Deadline task;
@@ -9,6 +17,6 @@ public class DeadlineTaskCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         taskList.addTask(task);
-        storage.addLine("D`UM`" + task.description + "`" + task.by);
+        storage.addLine("D`UM`" + task.getDescription() + "`" + task.getBy());
     }
 }
