+++ src/main/java/jamal/command/DeleteCommand.java
@@ -1,3 +1,9 @@
+package jamal.command;
+
+import jamal.task.TaskList;
+import jamal.util.Storage;
+import jamal.ui.Ui;
+
 public class DeleteCommand extends Command {
 
     protected int idx;
