+++ src/main/java/jamal/command/EventTaskCommand.java
@@ -1,3 +1,11 @@
+package jamal.command;
+
+import jamal.task.Event;
+import jamal.task.Task;
+import jamal.task.TaskList;
+import jamal.util.Storage;
+import jamal.ui.Ui;
+
 public class EventTaskCommand extends Command {
 
     protected Event task;
@@ -9,6 +17,6 @@ public class EventTaskCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         taskList.addTask(task);
-        storage.addLine("E`UM`" + task.description + "`" + task.start + "`" + task.end);
+        storage.addLine("E`UM`" + task.getDescription() + "`" + task.getStart() + "`" + task.getEnd());
     }
 }
