+++ src/main/java/jamal/command/ExitCommand.java
@@ -1,4 +1,10 @@
-public class ExitCommand extends Command{
+package jamal.command;
+
+import jamal.task.TaskList;
+import jamal.util.Storage;
+import jamal.ui.Ui;
+
+public class ExitCommand extends Command {
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
