+++ src/main/java/jamal/command/ListCommand.java
@@ -1,0 +1,13 @@
+package jamal.command;
+
+import jamal.task.TaskList;
+import jamal.util.Storage;
+import jamal.ui.Ui;
+
+public class ListCommand extends Command {
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        taskList.list();
+    }
+}
