+++ src/main/java/jamal/command/ListOngoingCommand.java
@@ -1,0 +1,13 @@
+package jamal.command;
+
+import jamal.task.TaskList;
+import jamal.util.Storage;
+import jamal.ui.Ui;
+
+public class ListOngoingCommand extends Command {
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        taskList.listOngoing();
+    }
+}
