+++ src/main/java/jamal/command/ListOverdueCommand.java
@@ -1,0 +1,13 @@
+package jamal.command;
+
+import jamal.task.TaskList;
+import jamal.util.Storage;
+import jamal.ui.Ui;
+
+public class ListOverdueCommand extends Command {
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        taskList.listOverdue();
+    }
+}
