+++ src/main/java/jamal/command/MarkCommand.java
@@ -1,4 +1,10 @@
-public class MarkCommand extends Command{
+package jamal.command;
+
+import jamal.task.TaskList;
+import jamal.util.Storage;
+import jamal.ui.Ui;
+
+public class MarkCommand extends Command {
 
     protected int idx;
