+++ src/main/java/jamal/exception/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package jamal.exception;
+
 public class InvalidCommandException extends Exception{
 
     public InvalidCommandException() {
