+++ src/main/java/jamal/ui/Ui.java
@@ -1,3 +1,5 @@
+package jamal.ui;
+
 import java.util.Scanner;
 
 public class Ui {
