+++ src/main/java/jamal/util/FileWrite.java
@@ -1,3 +1,5 @@
+package jamal.util;
+
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
