+++ src/main/java/jamal/util/Storage.java
@@ -1,11 +1,11 @@
+package jamal.util;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
-import java.util.Scanner;
-import java.util.stream.Stream;
 
 public class Storage {
